chore: Refactoring analysis
