+++ src/testronaut/cli/commands/__init__.py
@@ -2,6 +2,7 @@
 Command modules for the testronaut CLI.
 """
 
-from testronaut.cli.commands import analyze, generate, report, verify
+from testronaut.cli.commands import analyze_commands as analyze
+from testronaut.cli.commands import generate, report, verify
 
 __all__ = ["analyze", "generate", "verify", "report"]
