+++ src/testronaut/cli/commands/verify.py
@@ -1,60 +1,23 @@
 """
-Command module for verifying test results.
+Commands for verifying test results.
 """
 
-from pathlib import Path
-from typing import Optional
-
 import typer
-from rich.console import Console
 
-console = Console()
-app = typer.Typer(name="verify")
+from testronaut.cli.common import console
 
-@app.callback(invoke_without_command=True)
-def verify(
-    ctx: typer.Context,
-    test_plan: Optional[Path] = typer.Option(
-        None,
-        "--test-plan",
-        "-t",
-        help="Path to test plan file",
+# Create verify app
+verify_app = typer.Typer(help="Verify test results")
+
+
+@verify_app.command("results")
+def verify_results(
+    results_dir: str = typer.Argument(..., help="Directory with test results"),
+    report_format: str = typer.Option(
+        "text", "--format", "-f", help="Report format (text, json, html)"
     ),
-    output_file: Optional[Path] = typer.Option(
-        None,
-        "--output",
-        "-o",
-        help="Path to output verification results file",
-    ),
-):
-    """
-    Execute tests and verify results against expectations.
-
-    This command executes the tests defined in a test plan and verifies
-    the results against the expected outcomes.
-    """
-    # Get global options
-    tool = ctx.parent.params.get("tool")
-    output_dir = ctx.parent.params.get("output_dir")
-
-    # If no test plan is specified, use the default location
-    if not test_plan:
-        test_plan = Path(output_dir) / f"{tool}_test_plan.json"
-
-    # If no output file is specified, use the default location
-    if not output_file:
-        output_file = Path(output_dir) / f"{tool}_verification_results.json"
-
-    if not test_plan.exists():
-        console.print(f"[bold red]Error:[/bold red] Test plan file not found: {test_plan}")
-        console.print("Run 'testronaut generate' first to generate a test plan.")
-        raise typer.Exit(code=1)
-
-    console.print(f"[bold green]Verifying[/bold green] test results using plan: {test_plan}")
-    console.print(f"Output will be saved to: {output_file}")
-
-    # This would call the actual verifier in a real implementation
-    # from testronaut.core.verifier import verify_test_results
-    # verify_test_results(test_plan, output_file, docker_image=ctx.parent.params.get("docker_image"))
-
-    console.print("[bold green]Verification complete![/bold green]")
+) -> int:
+    """Verify test results and generate a report."""
+    console.print(f"Verifying results in: [bold]{results_dir}[/bold]")
+    console.print("[yellow]This functionality is still being implemented[/yellow]")
+    return 0
